refactor dagger scripts
