+++ terraform/.terraform-version
@@ -1,0 +1,1 @@
+1.1.7
