--- terraform/resources/provisions/oidc/provider.tf
@@ -1,12 +1,0 @@
-terraform {
-  required_providers {
-    aws = {
-      source  = "hashicorp/aws"
-      version = ">= 4.55.0"
-    }
-  }
-}
-
-provider "aws" {
-  region              = "ap-northeast-1"
-}
