--- terraform/resources/provisions/oidc/terragrunt.hcl
@@ -1,3 +1,0 @@
-include "root" {
-  path = find_in_parent_folders()
-}
